Integrate publication summary
