+++ context/app/static/js/components/publications/AggsList/AggsList.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import Typography from '@mui/material/Typography';
 
 import BulletedList from 'js/shared-styles/lists/bulleted-lists/BulletedList';
 import BulletedListItem from 'js/shared-styles/lists/bulleted-lists/BulletedListItem';
@@ -13,17 +14,21 @@ function AggsList({ uuid, field, associatedCollectionUUID }) {
 
   const buckets = searchData?.aggregations?.[field]?.buckets;
 
-  if (buckets) {
-    return (
-      <BulletedList>
-        {buckets.map((agg) => (
-          <BulletedListItem key={agg.key}>{agg.key}</BulletedListItem>
-        ))}
-      </BulletedList>
-    );
+  if (!buckets) {
+    return null;
   }
 
-  return null;
+  if (buckets.length === 1) {
+    return <Typography>{buckets[0].key}</Typography>;
+  }
+
+  return (
+    <BulletedList>
+      {buckets.map((agg) => (
+        <BulletedListItem key={agg.key}>{agg.key}</BulletedListItem>
+      ))}
+    </BulletedList>
+  );
 }
 
 export default AggsList;
