--- context/app/static/js/components/publications/PublicationSummary/style.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const DateContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: ${({ theme }) => theme.spacing(6)};
-`;
